agentos_core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentos_core"
version = "0.1.0"
description = "Central ReAct reasoning engine for Agent OS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agentos_memory",
    "agentos_skills",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "psutil>=5.9.0",
    "ollama>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_core*", "lane_queue*", "sandbox*", "llm_router*", "security*"]

[project.scripts]
agent-os = "main:main"
